viz/python: split viz_bindings.cpp along the C++ dep DAG
The bindings file was a single ~500-line TU; it would grow
quadratically as ProjectionLayer / OverlayLayer / RGBD types ship.
Splitting one-source-per-sub-module keeps the per-file size bounded
and mirrors the C++ link order.

Layout:
  bind_helpers.hpp   shared inline helpers (typestr_for, shape_for,
                     make_array_interface for the cuda/array
                     interface protocol)
  bind_core.cpp      enums + plain types + VizBuffer + HostImage
  bind_layers.cpp    QuadLayer + Config + Placement
  bind_session.cpp   FrameInfo / FrameTimingStats / GpuFrameTiming +
                     VizSessionConfig + VizSession (add_quad_layer
                     factory returns a non-owning QuadLayer handle
                     registered in bind_layers)
  viz_bindings.cpp   PYBIND11_MODULE entry point — ~15 lines that
                     calls bind_core / bind_layers / bind_session in
                     DAG order (types referenced by later signatures
                     register first)

No behavior change. Python namespace stays flat: `isaacteleop.viz`
exposes the same symbols, same docstrings, same GIL-release
annotations on blocking calls. ctest -L viz_python: 3/3 pass.

clang-format + pre-commit clean on all new sources.
